project/nhyc/nhyc/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
distlib==0.3.0
Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
filelock==3.0.12
httplib2==0.18.1
Markdown==3.2.2
mysql-connector-python==8.0.20
mysqlclient==1.4.6
numpy==1.18.4
pandas==1.0.4
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
source==1.2.0
sqlparse==0.3.1
virtualenv==20.0.17
